requirements-dev.txt
absl-py==0.10.0
annoy==1.17.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
astunparse==1.6.3
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.4.0
django-filter==2.3.0
django-heroku==0.3.1
django-mysql==3.8.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
filelock==3.0.12
gast==0.2.2
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
httpie==2.2.0
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mysql-connector-python==8.0.20
mysqlclient==2.0.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.15.87
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
white==0.1.2
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
tensorflow==2.0.0
tensorflow-hub==0.9.0
